Use the debugger gem
The ruby-debug19 is discontinued. The new gem is fully compatile with the
newest rubies, and it can be installed really fast with no tricks.

The gem is restricted to the development and test group, so it will not be
loaded in production mode (where there is no need to use a debugger).