Refactor in db/seeds.rb
The db/seeds.rb has to be prepared to be loaded multiple times, so it has to
take care that there is no duplicated data.

The Group.find_or_create_by_name is used to create the group only when it is
not found. The instance is stored in the Groups hash to be used later.

We removed the use of `user.update_attribute`, since there is no reason to
store the instance (via `User.create`) and then modify the value. The new
code use `User.new.tap` to initialize everything and make only one INSERT
in the database.

There is no guarantee that the admin group has the id 1, so the code

  update_attribute(:group_id, 1)

is dangerous. If we reset the groups table, the new created admins group
will have a bigger value.

The example admin user is created only when there is no users, so we can
run `rake db:seed` in production with no risk. If we don't check if there are
any users in the database we could create a security hole.