Refactor to ApplicationController#is_admin
We add a AdminUserIsRequired exception, so the helper is_admin can be used
at any point in the controller, not only in a before_filter.

In the previous version, when the client can not see the page, it was
redirected. This is very annoying, because the client loses the current URL.

The response sent a 403 (error) status , instead of a 302 (valid).

If the user has no logged, the application should show a form, and then
redirects to the original request.